Feature/detail card

+++ src/app/components/detail/detail.component.ts
@@ -9,14 +9,14 @@ import { CoffeeService } from '../../services/coffee.service';
 import { MatDialog } from '@angular/material/dialog';
 import { DialogComponent } from '../dialog/dialog.component';
 import { CardComponent } from "../card/card.component";
-import { DialogFormDirective } from 'src/app/directives/dialog-form.directive';
+import { DialogDeleteComponent } from '../dialog-delete/dialog-delete.component';
 
 @Component({
-    selector: 'app-detail',
-    standalone: true,
-    templateUrl: './detail.component.html',
-    styleUrls: ['./detail.component.scss'],
-    imports: [CommonModule, MatButtonModule, CardComponent, DialogFormDirective, RouterLink]
+  selector: 'app-detail',
+  standalone: true,
+  templateUrl: './detail.component.html',
+  styleUrls: ['./detail.component.scss'],
+  imports: [CommonModule, MatButtonModule, CardComponent, RouterLink]
 })
 export class DetailComponent {
 
@@ -27,19 +27,27 @@ export class DetailComponent {
   constructor(private _coffeeService: CoffeeService, private _activatedRoute: ActivatedRoute, private _router: Router, private _matDialog: MatDialog) {
   }
 
-  isClicked: boolean = false
-
   onUpdate(coffee: CoffeeModel) {
     this._matDialog.open(
       DialogComponent, {
-        data: coffee
-      }
+      data: coffee
+    }
     )
   }
 
   onDelete(id: number) {
-    this._coffeeService.delete(id).subscribe(() => {
-      this._router.navigate([''])
-    })
+    const dialogRef = this._matDialog.open(
+      DialogDeleteComponent)
+
+    dialogRef
+      .afterClosed()
+      .subscribe(result => {
+        if (result) {
+          this._coffeeService.delete(id).subscribe(() => {
+            this._router.navigate([''])
+          })
+        }
+      });
   }
 }
+
